fhe_client/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
concrete-ml>=1.9.0
scikit-learn>=1.3.0
joblib>=1.3
numpy>=1.24
